requirements.txt
streamlit>=1.31.1
streamlit-quill>=0.0.5
gspread>=5.12.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.122.0
html2text>=2020.1.16
